RF: move search view to disclosure; limit to cities with disclosure data.
