+++ disclosure/urls.py
@@ -8,19 +8,17 @@ from . import views
 
 admin.autodiscover()
 
-urlpatterns = patterns('',
-                       url(r'^$', views.homepage_view),
-                       url(r'^admin/', include(admin.site.urls)),
-                       url(r'^static/(?P<path>.*)$',
-                           'django.views.static.serve', {
-                               'document_root': settings.STATIC_ROOT,
-                               'show_indexes': True,
-                           }),
-                       url(r'', include('locality.urls')),
-                       url(r'^locations/(?P<fips_id>[0-9]+)$',
-                           views.location_view),
-                       url(r'^docs/', include('rest_framework_swagger.urls'))
-                       )
+urlpatterns = patterns(
+    '',
+    url(r'^$', views.homepage_view),
+    url(r'^admin/', include(admin.site.urls)),
+    url(r'^static/(?P<path>.*)$', 'django.views.static.serve', {
+        'document_root': settings.STATIC_ROOT,
+        'show_indexes': True}),
+    url(r'^search/', views.search_view),
+    url(r'^locations/(?P<fips_id>[0-9]+)$', views.location_view),
+    url(r'^docs/', include('rest_framework_swagger.urls'))
+)
 
 # Register all API Viewsets:
 api = SimpleRouter()
